Refactor handler function
